Fixed Ruby 3.1 synthax issue in Treetop generated parser
With Ruby 2.7 was deprecated the usage of `_1` (`_<n>`) identifier for future reseved usage.
With Ruby 3.0 the new synthax for "numbered parameters" was used, avoiding the usage of `_n` as method names.

Treetop's generated grammar contained `def _n` methods, mainly implied to the usage of `_` symbol in LSCL grammar.

This commit rename the `_` rule in the grammar to `cs` and replaces everywhere it's needed.

Explanation at https://github.com/cjheath/treetop/issues/49