mention the elasticsearch host variable for logstash docker container
The information on how to configure the setting `monitoring.elasticsearch.hosts` via an environment variable for a docker container is hard or nowhere to find. I would appreciate having this information mentioned. Alternatively a link to the place where the XPACK_MONITORING_ELASTICSEARCH_HOST setting/variable is documented would be great.