Fix wildcard pipeline IDs version error
The docs state that wildcards in the pipeline IDs are allowable from version 7.10 onwards. This is incorrect a the code to do this was only added in 7.11. This can be seen if you compare the files in question between the 7.10 and 7.11 branches - the functinon get_wildcard_pipelines does not exist in 7.10.

7.10 => https://github.com/elastic/logstash/blob/4344a874f6ace5e6bfb777cab46a93155eec7a22/x-pack/lib/config_management/elasticsearch_source.rb#L222

7.11 => https://github.com/elastic/logstash/blob/866633f2c3d230ceddbaf2b400753c95047d663b/x-pack/lib/config_management/elasticsearch_source.rb#L224