Fixes pure Java plugins buildscripts
With PR #14952 the version of libraries used in Logstash Gradle's buildscript was centralized.
However this causes problem in building pure Java plugin, exposed in #14963. The plugin's Gradle script apply the Logstash's core
rubyUtils.groovy which needs the definition for the versions. Before apply only plugins and buildscript keywords can be used and there
isn't an easy way to share buildscript's pluging version in projects that shares in such way unless we create published plugin.