[ci] Commit DRA -SNAPSHOT buildkite pipeline
This commit adds support for building + publishing DRA (-SNAPSHOT for
now) artifacts for Logstash. It builds on top of #15312 and therefore
only targets the `main` branch and is intended to be run manually
during a trial period before we retire the corresponding Jenkins job.

The structure is similar to Jenkins:

1. Three steps runs in parallel to build packages, x86_64 docker and
aarch64 docker artifacts.

2. Once 1. is successfully done, use release manager to publish the
artifacts.

We generate the pipeline steps for 1. and 2. dynamically (with a simple
Python script) to avoid repetition for future PRs: we will add a new
pipeline in a follow up PR to orchestrate scheduled runs of this
pipeline for the required branches.

Relates: https://github.com/elastic/ingest-dev/issues/1720