[DRA] Remove wait step in buildkite
DRA artifact builds support two optional parameters
`VERSION_QUALIFIER_OPT` and `DRA_DRY_RUN`.
The most important is `VERSION_QUALIFIER_OPT` which should be provided
when `alpha1` or similar versions need to be built.

Currently, after clicking new build, the pipeline takes ~20s to assemble
the steps and pauses the job waiting for these options to be filled
(or just accept the empty defaults) and press continue.

I feel that this could be trappy behavior because the majority of the
use cases don't need it, and it's likely that a user manually clicks
build and forgets that they'd need to confirm these parameters later on,
left with a hanging build.

This commit makes the parameters optional. If needed, they should be
defined explicitly in the New Build prompt, after expanding the
Options section.

The downside of this approach is that when needed, users need to
consult the documentation about the environment variable names.