[ci] Remove hardcoded branches from DRA schedule
This commit is the first making use of #15627 to remove hard coded
branches for the DRA Snapshot build schedule.

All the work here shouldn't be backported to other branches; the only
maintenance going forwards will be to keep `ci/branches.json` up to
date, as versions evolve.

Once this is verified working, we'll add a corresponding schedule
pipeline (in `catalog.info`) for the JDK matrix job.

Relates: https://github.com/elastic/ingest-dev/issues/2664