Remove temporary dependencies in logstash-core.gemspec
When the initial 8.11.0 release was created, a late downgrade in jruby from `9.4.3.0`
to `9.4.2.0` was required to avoid a silent crash during `Concurrent::TimerTask`. However, as part
of the downgrade,  we also added some cherry-picked dependencies to logstash-core to still
include some of the stlib fixes from `9.4.3.0` in the Logstash release.

We recently upgraded the version of jruby from `9.4.2.0` to `9.4.5.0`, but did not remove these
cherry-picked dependencies, some of which are now outdated, leading to issues such as #15687.

This commit removes those cherry-picked dependencies and updates the lockfile