[ci] No Slack alerts for pipeline scheduling job
As a follow up to #15705, this commit shushes Slack alerts for the
common Buildkite scheduling pipeline. This is because the pipeline
scheduler just triggers other pipelines which have their own dedicated
alerts setup, therefore we don't want to avoid duplicate alerts when
there is a failure with one of the triggered pipelines.