Run BK exhaustive pipeline when code is pushed
As we are close to sunsetting Jenkins by completing the
exhaustive tests Buildkite pipeline, it makes sense to start
already running the exhisting phases (testing and compatibility).

This commit defines the trigger to be code events, i.e. direct pushes
and merge commits, similar to what we currently have in Jenkins.