Backport of env2yaml arm64 Docker build context fixes
This commit contains fixes made to the main 8.x branch to generate
arm64 and amd64 env2yaml binaries, copy them into the build context
and correctly identify the correct binary in the Dockerfile to
include in the Docker image.

A clean backport was not available due to the change in format of
Dockerfile template from j2 to erb in #15142

This contains fixes from #15980 and #16053.