Ensure plugin config marked :deprecated logs to deprecation logger (#16863)
* Ensure plugin config marked :deprecated logs to deprecation logger (#16833)

Previously when the `:deprecated` modifier was used in the plugin config DSL a
log message was sent at `:warn` level to the main logger. This commit updates
that message to be routed *only* to the deprecation logger.

* Ensure plugin config marked `:deprecated` logs to deprecation logger

Previously when the `:deprecated` modifier was used in the plugin config DSL a
log message was sent at `:warn` level ONLY to the main logger. This commit updates
that message to be routed *both* to the deprecation logger as well as the main
logger. In 9.x this will *only* go to the deprecation logger.