First pass at a reprodicible env for running under FIPS mode
This commit adds a dockerfile intended to serve as a portable environment for
exploring running logstash in a FIPS compliant manner. The iteration at this
commit focuses on running *only* the tests with the BC provider and trust
stores. The same dockerfile has some commented out and unused config for doing
manual exploration with LS. Eventually this will be split out into separate
concerns.