Attempt to figure out timing
The major issue i'm running in to now is where exactly to make the tests use the
BC providers. Too early, and I run in to issues not being able to download
dependencies but too late and the tests will fail due to not being able to load.
This iteration looks at moving away from a script wrapping tasks  and instead moving it to the
build.gradle in the tasks directly. At this point its actually regressed in the
java tests. My current thinking is that by not explicitly adding the cached
locations to the classpath i've broken something.