Pattern for testing using standarized JVM config
In a previous iteration jvm options were attempted to be configured at the
gradle level. This presented a challeng in rspec loading because webmock was not
being loaded correctly with the fips providers. That initial approach was also
problematic because it introduced a fork in configuration changes between
configuring the JVM via the enviornment variables and files for running logstash
in the container vs in the tests.

This new apprach attempts to separate aout all of the test setup and building in
gradle from a pure "just run the tests" task. The fundamental idea is that we
dont want to use FIPS mode for downloading dependencies and building/preparing
an environment we only want that configured at the very end. This apprach
accomplishes that by teasing out the dependencies that trigger downloads etc in
gradel from tasks that will only run the tests. The dockerfile order will call
all the gradle tasks for setup, then configure FIPS mode and call the tests that
should be run under fips mode.